[REF] web: move to web the generatePdfThumbnail method
In this commit we move from mail to web the generatePdfThumbnail method.

task-5184941
